demos/video_editor/pipeline/requirements.txt
# Core dependencies
numpy>=1.24.0
soundfile>=0.12.0

# TTS
qwen_tts>=0.1.0
torch>=2.0.0
transformers>=4.30.0

# Audio transcription
faster-whisper>=0.9.0

# Google AI (Veo + Imagen)
google-genai>=0.4.0
google-cloud-storage>=2.10.0

# Testing
pytest>=7.0.0
